+++ src/driver/I2CDevice.hpp
@@ -13,11 +13,18 @@ namespace be::driver {
         be::I2C * i2c = nullptr ;
         uint8_t addr = 0 ;
 
+        uint8_t regSize = 1 ;
+        uint8_t regAddrSize = 1 ;
+
     public:
-        I2CDevice(JSContext * ctx, JSValue _jsobj=JS_NULL, uint8_t addr=0) ;
+        I2CDevice(JSContext * ctx, JSValue jsobj=JS_NULL, uint8_t regSize=1, uint8_t regAddrSize=1, uint8_t addr=0) ;
+        static JSValue constructor(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) ;
 
         virtual int setup() ;
         static JSValue setup(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) ;
-
+        
+        static JSValue readReg(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) ;
+        static JSValue writeReg(JSContext *ctx, JSValueConst this_val, int argc, JSValueConst *argv) ;
+        
     } ;
 }